sevici_api/requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
idna==3.4
importlib-metadata==6.6.0
Markdown==3.4.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
zipp==3.15.0
